+++ infra/envs/dev/outputs.tf
@@ -1,0 +1,24 @@
+output "vpc_id" {
+  description = "ID of the VPC"
+  value       = module.network.vpc_id
+}
+
+output "public_subnet_ids" {
+  description = "IDs of the public subnets"
+  value       = module.network.public_subnet_ids
+}
+
+output "private_subnet_ids" {
+  description = "IDs of the private subnets"
+  value       = module.network.private_subnet_ids
+}
+
+output "public_route_table_id" {
+  description = "ID of the public route table"
+  value       = module.network.public_route_table_id
+}
+
+output "internet_gateway_id" {
+  description = "ID of the Internet Gateway"
+  value       = module.network.internet_gateway_id
+}